requirements.txt
# Web Framework
flask
flask-cors

# AWS SDK
boto3
botocore

# AWS IoT SDK
AWSIoTPythonSDK

# Database Clients
pymongo
python-dotenv

# JWT and Cognito Auth
python-jose
requests

# Data Processing
pandas
numpy
scikit-learn
statsmodels
tensorflow
# PDF and Reports
matplotlib
reportlab>=3.6.0

# Testing
pytest
pytest-cov
unittest2

# For mocking in tests
mock

# Other utilities
python-dateutil